revert moving scene manager into micro
rationale: the scene manager has menu  and stats callbacks, which aren't in the App trait. the menu callback is a game-specific thing that just happens to be shared among multiple games. if i really want to make the scene manager a standard thing, i should standardize the menu callback first. same with the stats callback