add scene manager crate (copied from aetherbeats)
i'm back and forth on whether there should be a scene manager crate at all. the main downside is that the types have to be generic over Globals. that's a bummer, but having the scene management code copied into every project is also a bummer. idk which bummer is worse. mmmhhhh